bb_platform/requirements.txt
torch==1.6.0
torchvision==0.7.0
captum==0.2.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
cycler==0.10.0
Django==3.1
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
inflection==0.5.0
isort==5.5.2
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.1
opencv-python==4.2.0.34
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1


